### 2022-10-28 #4887 @editorialbot
[PRE REVIEW]: PyView: A general purpose tool for analyzing calcium imaging data
**Submitting author:** @ajkswamy (<a href="http://orcid.org/0000-0001-7928-877X">Ajayrama Kumaraswamy</a>)
**Repository:** https://github.com/galizia-lab/pyview
**Branch with paper.md** (empty if default branch): joss
**Version:** 1.1
