+++ view/python_core/io.py
@@ -112,7 +112,7 @@ class MultiTiffReaderInga():
     Input: .txt file with metadata
     '''
 
-    def __init__(self, txt_file: str):
+    def __init__(self, txt_file: str, measu: int):
     #def read_metadata_txt_file(fle):
         '''
         For multifile tif, Inga writes an info file as such:
@@ -151,6 +151,8 @@ class MultiTiffReaderInga():
         self.meta_dict = meta_dict
         self.data_path = pl.Path(txt_file).parent
         self.data_txt = txt_file
+        self.measu = measu
+        self.all_meta_data = pd.DataFrame()
 
 
 
@@ -161,8 +163,8 @@ class MultiTiffReaderInga():
         """
         # prepare the information: which measurements do I have?
         measurements = self.meta_dict['Stimuli']
-        measu = [t.strip() for t in measurements.split(',')] # now contains list of stimuli
-        measu_num = len(measu)
+        measurements_list = [t.strip() for t in measurements.split(',')] # now contains list of stimuli
+        measu_num = len(measurements_list)
         # list all the .tif files in the directory
         tif_files = list(self.data_path.glob('*.tif*'))
         tif_files.sort() # 
@@ -175,13 +177,13 @@ class MultiTiffReaderInga():
         measurementtime_first = None
 
         # iterate all measurements
-        for stim_ind, stimulus in enumerate(measu):
+        for stim_ind, stimulus in enumerate(measurements_list):
             # load first tif image to get metadata in tif file
-            tif_ind = stim_ind * int(self.meta_dict['Duration']) #!!!!!!!!!!!!!!wrong - files are not sorted
-            tif_file = tif_files[tif_ind+1]
+            tif_ind = stim_ind * int(self.meta_dict['Duration']) #assumes sorted files
+            tif_file = tif_files[tif_ind]
             
             # single information from TIF fil
-            single_metadata = pd.Series(dtype = 'object')
+            single_metadata = pd.Series(self.meta_dict, dtype = 'object') # all info from Inga is now also here
             with tifffile.TiffFile(tif_file) as tif:
                 single_metadata["FrameSizeX"] = tif.pages[0].tags['ImageWidth'].value
                 single_metadata["FrameSizeY"] = tif.pages[0].tags['ImageLength'].value
@@ -193,61 +195,76 @@ class MultiTiffReaderInga():
                 # results in, for example '=1635500442\n'
                 Meta_Imagetime = re.sub(r'[^\d.]+', '', Meta_Imagetime) # remove non-numeric characters
                 # unfortunately, this time does not match. Take file time instead from self.data_txt
-                ti_c = os.path.getctime(self.data_txt) # created
-                ti_m = os.path.getmtime(self.data_txt) # modified
+                ti_c = os.path.getctime(tif_file) # created
+                ti_m = os.path.getmtime(tif_file) # modified
                 measurementtime = np.min([ti_c, ti_m])
                 # there is something strange here: creation time is later than modification time in some cases. Take the earlier one               
                 single_metadata["UTC"] = measurementtime
-                # change format with  datetime.datetime.utcfromtimestamp(measurementtime)
+                # time is thought to be linear, i.e. measurements are in temporal order.
+                # therefore, we can calculate time within the experiment here
+                if not stim_ind: #value 0 for the first loop
+                    measurementtime_first = measurementtime
+                # now calculate relative time of this measurements, with respect to the first measurement
+                # get time from first measurement in this block
+                MTime = measurementtime - measurementtime_first
[…]
