+++ view/python_core/p1_class/__init__.py
@@ -4,9 +4,9 @@ from .filters import apply_filter
 from view.python_core.bleach_corr import get_bleach_compensator
 from view.python_core.background import get_background_frames
 from view.python_core.areas import get_area_for_p1, get_area_for_bleach_correction
-from view.python_core.measurement_list.importers import LSMImporter
+from view.python_core.measurement_list.importers import LSMImporter, IngaTif_Importer
 from view.python_core.foto import calc_foto1
-from view.python_core.io import load_pst, read_lsm, read_tif_2Dor3D, read_single_file_fura_tif, read_lif, read_SingleWavelengthTif_MultiFile
+from view.python_core.io import load_pst, read_lsm, read_tif_2Dor3D, read_single_file_fura_tif, read_lif, read_SingleWavelengthTif_MultiFileInga
 from view.python_core.paths import get_existing_raw_data_filename
 from view.python_core.stimuli import PulsedStimuliiHandler
 from view.python_core.calc_methods import get_calc_method
@@ -260,10 +260,12 @@ class P1SingleWavelengthAbstract(ABC):
             filename = get_existing_raw_data_filename(flags=flags, dbb=metadata.dbb1, extensions=current_extensions)
             if pl.Path(filename).suffix in current_extensions:
                 logging.getLogger("VIEW").info(f"(read_data_with_defaulting 1) Reading raw data from {filename}")
-                if '.lif' in current_extensions:
+                if ('.txt' in current_extensions) or ('.lif' in current_extensions):
                     #a .lif file containse several measurement, read_data for measu only
+                    #similarly a .txt file in INGA multi tiff format
                     return filename, self.read_data(filename, flags.flags['STG_Measu'])
                 else:
+                    logging.getLogger("VIEW").info(f"(read_data_with_defaulting 1) Cannot read raw data from {filename}")
                     return filename, self.read_data(filename)
             else:
                 raise FileNotFoundError()
@@ -277,6 +279,7 @@ class P1SingleWavelengthAbstract(ABC):
                     data, _ = read_tif_2Dor3D(filename)
                     return filename, data
                 else:
+                    logging.getLogger("VIEW").info(f"(read_data_with_defaulting 2) Cannot read raw data from {filename}")
                     raise FileNotFoundError()
             except FileNotFoundError as fnfe:
                 raise FileNotFoundError(f"{repr(fnfe)}.\n "
@@ -354,7 +357,7 @@ class P1SingleWavelengthTIF(P1SingleWavelengthAbstract):
         return data
 
 
-class P1SingleWavelength_multiTIF(P1SingleWavelengthAbstract):
+class P1SingleWavelength_multiTIFInga(P1SingleWavelengthAbstract):
     '''
     Measurement XYT is stored in separate Tiff file for each frame/timepoint
     '''
@@ -368,16 +371,46 @@ class P1SingleWavelength_multiTIF(P1SingleWavelengthAbstract):
         list of allowed file extensions. E.g.: [".tif"]
         :rtype: list
         """
-        return [".tif"]
+        return [".txt"] #this is the filename of Inga's information about the data
     
-    def read_data(self, filename_list: list):
+    def read_data(self, filename: str, measu: int):
         """
         read and return data in <filename_list>. Data is expected to be a numpy.ndarray of format XYT
         :param list filename_list: absolute path of raw data files for each frame on file system
         :rtype: numpy.ndarray
         """
-        data, _ = read_SingleWavelengthTif_MultiFile(filename_list)
+        if measu is None:
+            measu = 0 # default to the first measurement
[…]
