+++ view/python_core/measurement_list/importers.py
@@ -11,6 +11,7 @@ import datetime
 from abc import ABC, abstractmethod
 import xml.etree.ElementTree as ET
 from view.python_core.io import LIFReaderGio
+from view.python_core.io import MultiTiffReaderInga
 
 
 def calculate_dt_from_timing_ms(timing_ms: str) -> float:
@@ -416,215 +417,85 @@ class LSMImporter(BaseImporter):
 
 
 
-class P1SingleWavelengthTIFMultiFileImporter(BaseImporter):
+class IngaTif_Importer(BaseImporter):
     # added Sept2022 to import single wavelength stored in separate tif files for each frame
-    # data from Inga/Einat
+    # accompanied by a .txt file that structures the experiment
+    # data from Inga Petelski / Einat Couzin Fuchs
+    #
 
     def __init__(self, default_values: typing.Mapping):
 
         super().__init__(default_values)
-        self.associate_file_type = "Single Wavelength multi-frame Tif files"  # short text describing raw data files
+        self.associate_file_type = "Single Wavelength multi-frame Tif files with .txt info"  # short text describing raw data files
         self.associated_extensions = [".txt"]  # possible extensions of files containing metadata
         self.movie_data_extensions = [".tif"]  # possible extension of file containing data (calcium imaging movies)
         self.LE_loadExp = 32  # associated value of the flag LE_loadExp
 
 
-    def convert_metadata_to_lst_row(self, measu, fle, meta_info, default_row):
-        """
-        Convert values from meta_info to .lst nomenclature
-        :param meta_info['PsSzX']: dict, like the one returned by tifffile.TiffFile.lsm_metadata
-        :param default_row: pandas.Series, with default values
-        :return: pandas.Series
-        """
-
-        lst_line = default_row.copy()
-        lst_line["Label"] = meta_info['Label']
-        # converting from seconds to milliseconds
-        lst_line["Cycle"] = meta_info['GDMfreq'] 
-        lst_line["Lambda"] = meta_info['Lambda'] 
-        lst_line['UTC'] = meta_info['UTCTime'] 
-        # convert from meters to micrometers
-        lst_line["PxSzX"] = meta_info['PsSzX']
-        lst_line["PxSzY"] = meta_info['PsSzY']
-
-        analyze, dbb1_relative = self.get_path_relative_to_data_dir(fle)
-        lst_line["DBB1"] = meta_info['dbb1']
-        lst_line["dbb2"] = meta_info['dbb2']
-        lst_line["Analyze"] = analyze
-        lst_line["Measu"] = measu
-
-#additional info
-        lst_line["ExposureTime_ms"] = meta_info['ExposureTime_ms']
-        lst_line["AcquisitionDate"] = meta_info['AcquisitionDate']
-        lst_line["Binning"] = meta_info['Binning']
-        lst_line["StartTime"] = meta_info['StartTime']
-
-
-        return pd.DataFrame(lst_line).T
 
     def parse_metadata(self, fle: str, fle_ind: int,
                        measurement_filter: typing.Callable[[pd.Series], bool] = True) -> pd.DataFrame:
-        # load metadata from a .txt file, format defined by Inga, 2022
-        
-        
-        meta_dict = read_metadata_txt_file(fle)
-        # contains plenty of metadata
+        # load metadata from a .txt file, format defined by Inga Petelski, 2022
 
-        
-        
-        tif_file=pl.Path(fle)
-        with tifffile.TiffFile(tif_file) as tif:
-                metadata   = tif.imagej_metadata
-                # imagej_metadata does not work any more or never worked on stack - read metadata from first frame
-                if metadata is None:
-                    metadata = tif.pages[0].description
-    
[…]
