### 2022-11-13 #4887 @csoneson
👋🏻 @paxtonfitzpatrick, @XinhuiLi, @hvgazula - would two of you be interested in reviewing this submission for JOSS?

> PyView: A general purpose tool for analyzing calcium imaging data
