### 2022-11-13 #4887 @XinhuiLi
Hi @csoneson, I am afraid that I am not the best person to review this paper because I don't have any experience in calcium imaging. My friend Yannan Chen (@Yannan-Chen) has expertise in software development and calcium imaging data analysis, and he is willing to review this article. What is the deadline to submit the review?
