### 2022-11-14 #4887 @paxtonfitzpatrick
@csoneson I'll also gracefully decline this one. Calcium imaging is a bit outside my wheelhouse, and I won't have the bandwidth over the next few weeks to offer a quality review.
