### 2022-11-14 #4887 @hvgazula
Hello @csoneson, Thanks for the invite. But calcium imaging is not my domain of expertise, so I won't be able to give my best. I apologize for the inconvenience.

> 👋🏻 @paxtonfitzpatrick, @XinhuiLi, @hvgazula - would two of you be interested in reviewing this submission for JOSS?
>
> > PyView: A general purpose tool for analyzing calcium imaging data
