### 2022-11-14 #4887 @csoneson
@XinhuiLi, @paxtonfitzpatrick, @hvgazula - no problem, thanks for letting us know!

@Yannan-Chen, thanks for volunteering to review! You can find more information about the JOSS review process [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). We typically ask reviewers to provide an initial assessment within ~two weeks (from the start of review, which will happen as soon as we have secured two reviewers). However, we also note that it is often easier to post comments for the authors as you go through the submission, rather than waiting and posting them all at once.
