### 2022-11-14 #4887 @csoneson
👋🏻 @TomDonoghue, @adamltyson - would you be interested in reviewing this submission for JOSS?

> PyView: A general purpose tool for analyzing calcium imaging data
