### 2022-11-15 #4936 @editorialbot
[REVIEW]: PyView: A general purpose tool for analyzing calcium imaging data
**Submitting author:** @ajkswamy (<a href="http://orcid.org/0000-0001-7928-877X">Ajayrama Kumaraswamy</a>)
**Repository:** https://github.com/galizia-lab/pyview
**Branch with paper.md** (empty if default branch): joss
**Version:** 1.3
**Archive:** 10.12751/g-node.ezh5w2
