### 2022-11-16 galizia-lab/pyview#18 @adamltyson
Installation instructions for Apple Silicon
Hi :wave:,

I'm [reviewing the PyView JOSS paper](https://github.com/openjournals/joss-reviews/issues/4936).

I'm having trouble installing PyView on my machine which I assume is partially due to Apple Silicon M1 chip. It fails very early (e.g. Python 3.7 is not available for M1 on conda forge).

Could you add instructions for installing the package on newer macs, or better yet, an environment file?

Thanks!
