### 2022-11-16 galizia-lab/pyview#18 @galizia
Hi!

Thanks for the question -
I am trying to install conda right now, since yesterday I have an M2 mac, and it does not work yet. Maybe in a few days I will be able to help you. Mac has really changed many things, Anaconda is impossible to install, conda seems to work from the comand line:

https://repo.anaconda.com/archive/Anaconda3-2022.10-MacOSX-arm64.sh <https://repo.anaconda.com/archive/Anaconda3-2022.10-MacOSX-arm64.sh>

There is an environment file that can be used in conda, and works well with my previous mac. I did not yet have the time to test it with the M2 mac - sorry about that. I will try to do that asap, but I am travelling at the moment and cannot promise how fast I’ll be.

I’m copying Ajayrama Kumaraswamy: he is the computer scientist in our team (I am the biologist), but he does not run a mac, so he might be helpful or might not.

Best
Giovanni

*******************
Prof. Dr. Giovanni Galizia
Direktor Zukunftskolleg
Lehrstuhl Neurobiologie
Permanent Fellow Wissenschaftskolleg zu Berlin

***@***.***
http://neuro.uni.kn
*****************************

> On 16. Nov 2022, at 19:02, Adam Tyson ***@***.***> wrote:
>
>
> Hi 👋,
>
> I'm reviewing the PyView JOSS paper <https://github.com/openjournals/joss-reviews/issues/4936>.
>
> I'm having trouble installing PyView on my machine which I assume is partially due to Apple Silicon M1 chip. It fails very early (e.g. Python 3.7 is not available for M1 on conda forge).
>
> Could you add instructions for installing the package on newer macs, or better yet, an environment file?
>
> Thanks!
>
