### 2022-11-17 galizia-lab/pyview#18 @adamltyson
Thanks @galizia. Let me know if I can be of any help. We have quite a bit of Python/PyQt5 software that was troublesome to get working on macOS with Apple Silicone, but it is possible.
