### 2022-11-18 galizia-lab/pyview#18 @galizia
Hi again!

I have now replicated the python=3.7 problem, python=3.8 seems to work, could you try that?
(I cannot check that yet, because I am having problems installing ffmpeg right now - as soon as I solved that, I will tell you. Therefore, if you prefer to wait until I have done that, no problem)
(Python=3.10 will need a few adjustments, since scipy has changed quite a bit).

Thanks,
Giovanni

*******************
Prof. Dr. Giovanni Galizia
Direktor Zukunftskolleg
Lehrstuhl Neurobiologie
Permanent Fellow Wissenschaftskolleg zu Berlin

***@***.***
http://neuro.uni.kn
*****************************

> On 16. Nov 2022, at 19:02, Adam Tyson ***@***.***> wrote:
>
>
> Hi 👋,
>
> I'm reviewing the PyView JOSS paper <https://github.com/openjournals/joss-reviews/issues/4936>.
>
> I'm having trouble installing PyView on my machine which I assume is partially due to Apple Silicon M1 chip. It fails very early (e.g. Python 3.7 is not available for M1 on conda forge).
>
> Could you add instructions for installing the package on newer macs, or better yet, an environment file?
>
> Thanks!
>
