### 2022-11-20 61f1ee0
changes to avoid future warning from pandas
