### 2022-11-20 6ea2be5
removed easygui from env file. will load in PyPI
