### 2022-11-20 galizia-lab/pyview#18 @galizia
Hi there!

Thanks for your patience!

We have now upgraded the code to python 3.10. You will find a new conda_env_development_py310.yml file to create a conda environment. On my silicon M2 it works fine!

Please let me know if you need any more information.

Best,
Giovanni

*******************
Prof. Dr. Giovanni Galizia
Direktor Zukunftskolleg
Lehrstuhl Neurobiologie
Permanent Fellow Wissenschaftskolleg zu Berlin

***@***.***
http://neuro.uni.kn
*****************************

> On 16. Nov 2022, at 19:02, Adam Tyson ***@***.***> wrote:
>
>
> Hi 👋,
>
> I'm reviewing the PyView JOSS paper <https://github.com/openjournals/joss-reviews/issues/4936>.
>
> I'm having trouble installing PyView on my machine which I assume is partially due to Apple Silicon M1 chip. It fails very early (e.g. Python 3.7 is not available for M1 on conda forge).
>
> Could you add instructions for installing the package on newer macs, or better yet, an environment file?
>
> Thanks!
>
