### 2022-11-21 galizia-lab/pyview#18 @ajkswamy
Hi @adamltyson :wave:

Thanks for raising the issue. I have made a new release (v1.2) incorporating all the changes mentioned above by @galizia . Please try installing this version. We are available if you have questions or issues.

Thanks!
