### 2022-11-21 #4936 @ajkswamy
Hello @csoneson @adamltyson @Yannan-Chen

Thanks for organizing the review processing and agreeing to review our paper!

A quick update: thanks to @adamltyson's [issue](https://github.com/galizia-lab/pyview/issues/18), we have now migrated PyView to Python 3.10 and made a new release. We request you to use this release when reviewing.

We are looking forward to your issues, comments, questions and reviews!

Thanks!
