+++ log2list_examples/create_measurement_list_Inga.py
@@ -1,0 +1,424 @@
+'''
+Creates .lst file from/for  Inga single-frame single-tif files
+Author: Giovanni, September 2022, based on template in VIEW folder 
+
+Expected data structure:
+    In the folder "01_DATA", each animal has a folder
+    - Data is from SetupB, new CCD as of 2021
+    - Format of the data is: every single frame is an individual .tif file, with sequential numbering
+    - Next to all .tif files, there is a .txt file with the experimental parameters
+    
+    There is a sister folder "02_LISTS" (created if not yet present)
+    Location is in STG_MotherOfAllFolders (set it below)
+    
+Output:
+    In the folder "02_LISTS", for every animal:
+        There will be a file Animal.lst.xlsx, e.g. "glom17_210923_bee11.lst.xlsx"
+        That file contains one line for each measurement.
+        Measurements that have no time dimension (snapshots, z-stacks) will have "0" in the column "analyze"
+
+Dataformat:
+    DBB1 contains the name of the .txt file
+    dbb2 contains a list with the individual .tif files, one for each frame
+    
+
+        
+What to do next?
+    In THIS file, change values that are global
+    or insert a function that can extract odor name or concentration name from somewhere
+    
+    In the Animal.lst.xlsx file, correct/complete entries (e.g. odor names, odor concentrations)
+    Make sure stimulus timing information is correct
+    
+    When you run this program again on the same dataset, and the Animal.lst.xlsx file is already present,
+    some columns will NOT be overwritten, but will be taken from the previous .lst.xlsx file,
+    protecting your manually entered information. 
+    Which columns? Specify them below in 
+    overwrite_old_values
+
+Why do I need a .lst.xlsx file?
+    Load measurements in pyVIEW using this .lst file, so that stimulus information is correct
+    For off-line analysis, information is taken from this file. 
+
+Good to know:
+    Information, where possible, is taken from the OME header of the incoming .tif file.
+    If that information is wrong, incomplete, or else, modify the code 
+    
+'''
+from view.python_core.measurement_list import MeasurementList
+from view.python_core.measurement_list.importers import get_importer_class
+from view.python_core.flags import FlagsManager
+from collections import OrderedDict
+import pandas as pd
+import logging
+import pathlib as pl
+import numpy as np
+from view.python_core.io import read_SingleWavelengthTif_MultiFileInga, write_tif_2Dor3D
+
+
+logging.basicConfig(level=logging.INFO)
+
+# ------------------ names of columns that will be overwritten by old values -------------------------------------------
+# -- if you run the same animal a second time!
+# ------ these will only be used if a measurement list file with the same name as current output file exists -----------
+
+overwrite_old_values = ["Line", "PxSzX", "PxSzY", "Age", "Sex", "Prefer",
+                        "Comment", "Analyze", "Odour", "OConc"]
+
+# ______________________________________________________________________________________________________________________
+
+
+
+
+# ------------------- Some parameters about experimental setup, data structure and output file type --------------------
+# 3 for single wavelength Till Photonics Measurements
+# 4 for two wavelength Till Photonics Measurements
+# 20 for Zeiss Confocal Measurements
+# 21 for Leica Confocal Measurements: .lif file
+# 32 for multiple TIF files, one for each frame
+LE_loadExp = 32 #32 for multiple TIF files
+
+# Mother of all Folders of your dataset
+# On Windows, if you copy paths from the file explorer, make sure the string below is always of the form r"......"
+STG_MotherOfAllFolders = r'/Users/galizia/Documents/DATA/inga_calcium/'
[…]
