### 2022-11-27 8ac9a6a
Modified animal naming for Inga to make it consistent with view
