+++ conda_env_development_py310.yml
@@ -1,20 +1,0 @@
-name: view_dev_py310
-channels:
-  - defaults
-  - conda-forge
-dependencies:
-  - python=3.10
-  - pip
-  - jinja2
-  - seaborn
-  - bokeh
-  - xlrd
-  - pyqtgraph
-  - whoosh
-  - moviepy  # only available on conda-forge
-  - openpyxl
-  - ipdb
-  - nose
-  - ipython
-  - jupyter
-  - appdirs
