### 2022-11-28 09c0f29
Inga tapestries: sort by stimulus name
