### 2022-12-05 galizia-lab/pyview#20 @adamltyson
Consider adding some sample datasets
Hi,

RE: [reviewing this paper for JOSS](https://github.com/openjournals/joss-reviews/issues/4936).

I've got the software running now on my machine. I might have missed a link somewhere, but I couldn't see a link to any sample data in the correct format/structure.

It would be very helpful to test the software, particularly for potential new users, if there was a link to sample data. Ideally this could be fetched from within the GUI.

Thanks,
Adam
