### 2022-12-05 galizia-lab/pyview#21 @adamltyson
Add a simple tutorial for new users
Hi,

RE: [reviewing this paper for JOSS](https://github.com/openjournals/joss-reviews/issues/4936).

The documentation is very comprehensive, but (at least to me) it seems like a technical reference for existing users. As a newcomer to the software, it's not immediately obvious where best to start.

Could there be a step by step walk-through to introduce the basic concepts?

Thanks,
Adam
