### 2022-12-05 galizia-lab/pyview#20 @ajkswamy
Hi Adam

I have added a section "Exploring PyView using Example datasets" to the page [Getting Started](https://github.com/galizia-lab/pyview/wiki/Getting-Started) of the wiki. Please download the example datasets and try our GUI/scripting workflows outlined on the same page. Let us know if you face any problems/issues!

Thanks!

Sincerely
Ajay
