### 2022-12-06 galizia-lab/pyview#20 @adamltyson
Thanks Ajay,

I'm trying to review this software from a users point of view, but even with this sample data it's hard to know where to begin. I think an introductory, step by step tutorial (see https://github.com/galizia-lab/pyview/issues/21) would be very useful.

Adam
