### 2022-12-07 #4936 @adamltyson
The paper itself looks good. It may be nice to cite the [suite2p pre-print](https://www.biorxiv.org/content/10.1101/061507v2) for consistency (e.g. the CaImAn paper is cited).
