### 2022-12-14 galizia-lab/pyview#29 @Yannan-Chen
Apply and inspect processing results after setting Flags
re: openjournals/joss-reviews#4936

Hi, is there a guideline about how to inspect the processing result after setting the Flags? E.g. after setting the filter options.

As in #21 , it will be great to have a demo with which walks through the pipeline step by step with screenshots/anticipated outcomes.
