### 2022-12-14 galizia-lab/pyview#29 @galizia
Dear Yannan Chen,

this is a good suggestion - I will work on that in the next few days.

Best,
Giovanni

> On 14. Dec 2022, at 02:57, Yannan Chen ***@***.***> wrote:
>
>
> re: openjournals/joss-reviews#4936 <https://github.com/openjournals/joss-reviews/issues/4936>
> Hi, is there a guideline about how to inspect the processing result after setting the Flags? E.g. after setting the filter options.
>
> As in #21 <https://github.com/galizia-lab/pyview/issues/21> , it will be great to have a demo with which walks through the pipeline step by step with screenshots/anticipated outcomes.
>
