### 2022-12-22 8a1ba3c
added LIFloader to direct loading in GUI; fixed bug with OME Fura single TIFF
