+++ synthetic_data/view_synthetic_666.yml
@@ -1,0 +1,109 @@
+CTVM_Method: 0
+CTV_FeatureNumber: 0
+CTV_Method: 22
+CTV_MethodFile: not set yet
+CTV_StimulusNumber: 0
+CTV_firstframe: 25
+CTV_lastframe: 45
+CTV_scalebar: false
+Data_Mean_Filter: 0
+Data_Mean_Filter_space: 3
+Data_Mean_Filter_time: 3
+Data_Median_Filter: 0
+Data_Median_Filter_space: 3
+Data_Median_Filter_time: 3
+Data_ReplaceInitFrames: 0
+GDM_chunkPostStim: 0.0
+GDM_chunkPreStim: 0.0
+GDM_outputType: full_traces
+GDM_withinArea: false
+LELog_ExcludeSeconds: 0
+LELog_InitialFactor: 1
+LE_BleachCorrMethod: None
+LE_BleachCutBorder: 20.0
+LE_BleachExcludeArea: true
+LE_BleachExcludeStimulus: true
+LE_BleachStartFrame: 2
+LE_CalcMethod: 3
+LE_DefaultBackgroundRange: (3, 5)
+LE_PrestimEndBackground: 1
+LE_ScatteredLightFactor: 0.0
+LE_ScatteredLightRadius: 50.0
+LE_StartBackground: 4
+LE_StimulusBasedBackground: true
+LE_labelColumns: ('Measu',)
+LE_loadExp: 666
+RM_ROIThreshold: 75.0
+RM_ROITrace: 3
+RM_Radius: 5
+RM_differentViews: false
+SO_MV_colortable: 12
+SO_MV_scalemax: 0.3
+SO_MV_scalemin: -0.08
+SO_Method: 0
+SO_bgColor: w
+SO_cutborder: 0
+SO_fgColor: k
+SO_fontName: None
+SO_indiScale3factor: 0.2
+SO_individualScale: 3
+SO_lowerThreshPositiveResps: a0
+SO_percentileScale: false
+SO_percentileValue: 0.0
+SO_reverseIt: false
+SO_rotateImage: 0
+SO_scaleLegendFactor: 1.0
+SO_showROIs: 0
+SO_thresholdOn: none
+SO_thresholdScale: full
+SO_thresholdShowImage: foto1
+SO_upperThreshNegativeResps: a0
+SO_withinArea: false
+SO_xgap: 30
+STG_Datapath: 01_DATA
+STG_Measu: 9
+STG_MotherOfAllFolders: '!!parent of this file will be used!!'
+STG_OdorAreaPath: 03_AREAS
+STG_OdorInfoPath: 02_LISTS
+STG_OdorReportFile: FakeData.report
+STG_OdorReportPath: 04_OUTPUT
+STG_OdormaskPath: 03_COOR
+STG_ProcessedDataPath: 05_REPORT
+STG_ReportTag: Synthetic_data_strip
+STG_TempArchivePath: not set yet
+Signal_FilterSpaceFlag: false
+Signal_FilterSpaceSize: 1.0
+Signal_FilterTimeFlag: false
+Signal_FilterTimeSize: 3.0
+VIEW_CorrSignals: true
+VIEW_InitCorr: 0
+VIEW_batchmode: false
+mv_FirstFrame: 0
+mv_LastFrame: -1
+mv_SpeedFactor: 1.0
+mv_bgColor: black
+mv_bitrate: 1024k
+mv_correctStimulusOnset: 0
+mv_cutborder: 0
+mv_displayTime: 1.0
+mv_exportFormat: libx264
+mv_fgColor: white
+mv_fontName: None
+mv_indiScale3factor: 0.0
+mv_individualScale: 3
[…]
