+++ README.md
@@ -14,5 +14,9 @@ If you find bugs, have questions about its usage, or have ideas for new features
 # Contributing
 Please consult [this page on the wiki](https://github.com/galizia-lab/pyview/wiki/Developer-Guide) for a Developer Guide. Pull requests corresponding to well documented issues in this repository are welcome!
 
+# Tutorial
+Copy the folder "synthetic_data" into a place of your system where you keep data. (If you installed as a developer, the folder is in the program tree. If you installed as a simple user, you should download the folder from the github location).
+Open the file VIEW_Tutorial.pdf. This will guide you through a hands-on quick tour using a synthetic data set, showing you how to use VIEW, ILTIS and off-line analysis. 
+
 # Testing
 TODO
