### 2023-01-08 b1598fa
added LE_LoadExp 665,667, 676 to direct load; minor bug fix
