### 2023-01-08 9e9f77e
added LE_LoadExp 665,667, 676 to direct load; minor bug fix
