### 2023-01-09 62da18c
split conda files into linux and mac+win to add glibc dependency
