+++ conda_py310_linux.yml
@@ -1,0 +1,21 @@
+name: view_py310
+channels:
+  - defaults
+  - conda-forge
+dependencies:
+  - python=3.10
+  - pip
+  - scipy
+  - scikit-image
+  - jinja2
+  - seaborn
+  - bokeh
+  - xlrd
+  - pyqtgraph
+  - whoosh
+  - openpyxl
+  - jupyter
+  - appdirs
+# only available on conda-forge
+  - moviepy
+  - libstdcxx-ng
