+++ conda_py310_dev_linux.yml
@@ -5,16 +5,20 @@ channels:
 dependencies:
   - python=3.10
   - pip
+  - scipy
+  - scikit-image
   - jinja2
   - seaborn
   - bokeh
   - xlrd
   - pyqtgraph
   - whoosh
-  - moviepy  # only available on conda-forge
   - openpyxl
-  - ipdb
   - nose
   - ipython
   - jupyter
   - appdirs
+# only available on conda-forge
+  - moviepy
+  - ipdb
+  - libstdcxx-ng
