### 2023-01-10 #4936 @adamltyson
Hey @ajkswamy, I saw that you added a tutorial PDF. The software is much, much easier to understand now. Considering how much work has gone into this software, and it's broad functionality, it would be a shame if it's not used by more people. I would recommend documenting everything as comprehensively as the tutorial, and make it as easy to find for new users as possible.

However, there are very few tools that couldn't do with a bit more documentation (including all of mine!), so considering the installation procedure is also much improved, I'm now happy to recommend acceptance! :tada: cc @csoneson.
