### 2023-01-10 1e88278
updated readme; setup points to ILTIS v1.6
