### 2023-01-15 #4936 @Yannan-Chen
Hi @csoneson ,
In this submission, the authors provide a software offering a wide range of features for working with calcium imaging data interactively. The tool is user-friendly and easy to install, with clear and straightforward documentation. Overall, I’m happy to recommend this tool to be accepted by JOSS!
