### 2023-01-15 e577a84
Added folder with Inga specific files within VIEWtemplates
