+++ VIEWtemplates/Inga/view_flags_for_tapestry.yml
@@ -1,0 +1,106 @@
+Data_Mean_Filter: 3
+Data_Mean_Filter_space: 3
+Data_Mean_Filter_time: 0
+Data_Median_Filter: 0
+Data_Median_Filter_space: 3
+Data_Median_Filter_time: 0
+CTVM_Method: 0
+CTV_Method: 22
+CTV_MethodFile: pyVIEW_PROGS/localCTV.py
+CTV_firstframe: 20
+CTV_lastframe: 30
+CTV_scalebar: true
+VIEW_CorrSignals: false
+Signal_FilterSpaceFlag: true
+Signal_FilterSpaceSize: 2.0
+Signal_FilterTimeFlag: false
+Signal_FilterTimeSize: 3.0
+LE_BleachCorrMethod: None
+LE_BleachCutBorder: 10.0
+LE_BleachExcludeArea: false
+LE_BleachExcludeStimulus: true
+LE_BleachStartFrame: 2
+LE_DefaultBackgroundRange: (3, 20)
+LELog_ExcludeSeconds: 3
+LELog_InitialFactor: 3
+LE_PrestimEndBackground: 0
+LE_ScatteredLightFactor: 0.0
+LE_ScatteredLightRadius: 50.0
+LE_StartBackground: 4
+LE_StimulusBasedBackground: true
+STG_MotherOfAllFolders: '!!parent of this file will be used!!'
+RM_NewColumn: true
+RM_NextPosition: (0, 0)
+RM_PlotTrace: false
+RM_ROIThreshold: 0.75
+RM_ROITrace: 0
+RM_Radius: 10
+RM_differentViews: false
+SO_Method: 0
+SO_bgColor: w
+SO_cutborder: 0
+SO_fgColor: k
+SO_fontName: None
+SO_indiScale3factor: 0.2
+SO_individualScale: 3
+SO_lowerThreshPositiveResps: a0
+SO_percentileScale: false
+SO_percentileValue: 0.0
+SO_reverseIt: false
+SO_rotateImage: 0
+SO_scaleLegendFactor: 1.0
+SO_showROIs: 0
+SO_thresholdOn: none
+SO_thresholdScale: full
+SO_thresholdShowImage: foto1
+SO_upperThreshNegativeResps: a0
+SO_withinArea: false
+SO_xgap: 30
+STG_Datapath: 01_DATA
+STG_Measu: 4
+STG_OdorAreaPath: 05_AREA
+STG_OdorInfoPath: 02_LISTS
+STG_OdorReportFile: outputfile
+STG_OdorReportPath: 06_OUTPUT
+STG_OdormaskPath: 05_COOR
+STG_ProcessedDataPath: Processed Data
+STG_ReportTag: report_tag
+STG_TempArchivePath: not set yet
+VIEW_InitCorr: -1
+VIEW_batchmode: false
+SO_MV_colortable: 13
+GDM_withinArea: false
+mv_FirstFrame: 1
+mv_LastFrame: -1
+mv_SpeedFactor: 1.0
+mv_bgColor: k
+mv_bitrate: 1024k
+mv_correctStimulusOnset: 0
+mv_cutborder: 0
+mv_displayTime: 0.8
+mv_exportFormat: libx264
+mv_fgColor: w
+mv_fontName: None
+mv_indiScale3factor: 0.2
+mv_individualScale: 3
+mv_lowerThreshPositiveResps: a0
+mv_markStimulus: 1
+mv_percentileScale: false
+mv_percentileValue: 0.0
+mv_reverseIt: true
+mv_rotateImage: 0
+mv_scaleLegendFactor: 100.0
+mv_showROIs: 1
[…]

[…]
