### 2023-01-22 #4936 @csoneson
Hi @ajkswamy - sorry for the delay. I had a quick look through the paper, and at the moment it is a bit too long for JOSS standards. In particular, see [here](https://joss.readthedocs.io/en/latest/submitting.html):
> The paper should be between 250-1000 words. Authors submitting papers significantly longer than 1000 words may be asked to reduce the length of their paper.

Your paper currently:
> Wordcount for paper.md is 2603

Can you see if you can bring the paper a bit closer to the suggested length? In particular some of the technical content could perhaps be moved to the documentation, and I think some of the non-technical sections can be made more concise without losing too much information. Thanks!
