### 2023-01-29 6250a81
Substitued arch. of data processing with link to wiki; trimmed statement of need
