+++ manuscript_joss/paper.md
@@ -34,18 +34,9 @@ Information about example workflows, galleries of examples outputs, and guides f
 
 # Statement of need
 
-Optical imaging experiments, including calcium imaging experiments, can have long durations, in particular if performed in vivo. Typically, they consist of many measurements with different experimental settings (stimulus type and intensity, pharmacological treatments, animal age or genotype, to name but a few). The resulting data needs pre-processing of various kinds, depending on the system that is being investigated (e.g. olfactory or visual system, or animal species), the activity reporter used (e.g. GCamP, FURA), and the optical properties of the setup (e.g. 2-photon imaging, widefield imaging; single wavelength or ratiometric imaging). Bleaching and baseline correction, movement correction, scattered light correction are necessary, selection of appropriate regions of interest, and exclusion or areas that are not neurons are among the most important steps here. Adjusting parameters for data treatment can be tedious, and necessitates interactive visualization by the researcher. 
-Conversely, statistical analysis across experimental animals, and across experimental parameters (e.g. different stimuli for sensory systems, or different animal groups), necessitate that all datasets are treated in an identical way. Therefore, data generation for statistical analysis should be done not using an interactive mode, but rather using batch processing, with equal parameter settings across all datasets. 
-PyView has been developed with the following primary aims in mind:
+Optical imaging experiments, including calcium imaging experiments, can have long durations, in particular if performed in vivo. Typically, they consist of many measurements with different experimental settings (stimulus type and intensity, pharmacological treatments, animal age or genotype, to name but a few). The resulting data needs pre-processing of various kinds, depending on the system that is being investigated (e.g. olfactory or visual system, or animal species), the activity reporter used (e.g. GCamP, FURA), and the optical properties of the setup (e.g. 2-photon imaging, widefield imaging; single wavelength or ratiometric imaging). Bleaching and baseline correction, movement correction, scattered light correction are necessary, selection of appropriate regions of interest, and exclusion or areas that are not neurons are among the most important steps here. Adjusting parameters for data treatment can be tedious, and necessitates interactive visualization and iterative experimentation by the researcher. 
 
- - Enable experimenters to easily, quickly and interactively assess the quality of calcium imaging data by visualizing raw data along with prospective estimates of neural activity inferred from the raw data. 
- - Interactively find the most suitable processing steps and parameters for removing artifacts in raw data and estimating activity levels. Save the steps and parameters to files in a standardized format (.yml files). 
- - Apply the same processing steps with the same parameters to all data acquired in a line of investigation in a non-interactive way using scripts. 
- - Create publication-ready visualizations: images and movies of individual recordings with many kinds of annotations, and HTMLs containing multiple images organized in a grid. 
- - Export the processed data to CSV or MAT files in a standardized format for further investigation with external tools. 
- - Avoid multi-tool, multi-step processing, by allowing to act as wrapper for other software packages. 
- - Interactively create regions-of-interest (ROIs) of diverse shapes and save them to files, which could be used later for identifying neural substrates, for specifying brain regions where processing should focus, and to export corresponding time traces. 
- - Organize raw data, metadata, files containing ROIs, ad-hoc images and all other files related to a line of investigation in a standardized folder structure, with clear separation of original data and processed/interpreted data.
[…]
