+++ manuscript_joss/paper.md
@@ -34,7 +34,7 @@ Information about example workflows, galleries of examples outputs, and guides f
 
 # Statement of need
 
-Optical imaging experiments, including calcium imaging experiments, can have long durations, in particular if performed in vivo. Typically, they consist of many measurements with different experimental settings (stimulus type and intensity, pharmacological treatments, animal age or genotype, to name but a few). The resulting data needs pre-processing of various kinds, depending on the system that is being investigated (e.g. olfactory or visual system, or animal species), the activity reporter used (e.g. GCamP, FURA), and the optical properties of the setup (e.g. 2-photon imaging, widefield imaging; single wavelength or ratiometric imaging). Bleaching and baseline correction, movement correction, scattered light correction are necessary, selection of appropriate regions of interest, and exclusion or areas that are not neurons are among the most important steps here. Adjusting parameters for data treatment can be tedious, and necessitates interactive visualization and iterative experimentation by the researcher. 
+Optical imaging experiments, including calcium imaging experiments, can have long durations, in particular if performed in vivo. Typically, they consist of many measurements with different experimental settings (stimulus type and intensity, pharmacological treatments, animal age or genotype, to name but a few). The resulting data needs pre-processing of various kinds, depending on the system that is being investigated (e.g. olfactory or visual system, or animal species), the activity reporter used (e.g. GCamP, FURA), and the optical properties of the setup (e.g. 2-photon imaging, widefield imaging; single wavelength or ratiometric imaging). Bleaching and baseline correction, movement correction, scattered light correction are necessary, selection of appropriate regions of interest, and exclusion of areas that are not neurons are among the most important steps here. Adjusting parameters for data treatment can be tedious, and necessitates interactive visualization and iterative experimentation by the researcher. 
 
 Conversely, statistical analysis across experimental animals, and across experimental parameters (e.g. different stimuli for sensory systems, or different animal groups), necessitate that all datasets are treated in an identical way. Therefore, data generation for statistical analysis should be done not using an interactive mode, but rather using batch processing, with identical parameter settings across all datasets and nonetheless using the same code as during GUI experimentation.
 	
@@ -53,7 +53,7 @@ This contains all the classes, functions and definitions that form the core of P
 This is a GUI written in PyQt5 with which users can use the functionalities implemented in PyView-core: configure parameters, save configuration files, load data, create representative images and movies, and export processed data into CSV files for further analysis. 
 
 ## ILTIS
-ILTIS (<https://github.com/grg2rsr/ILTIS>) is a stand-alone package used for interactive calcium imaging data visualization, a fork(<https://github.com/galizia-lab/ILTIS>) of which has been integrated into PyView-GUI. This allows interactive visualization of imaging data and inspecting individual image frames or pixelwise time traces. ILTIS can also be used to manually segment images and save the results into files.
+ILTIS (<https://github.com/grg2rsr/ILTIS>) is a stand-alone package used for interactive calcium imaging data visualization, a fork (<https://github.com/galizia-lab/ILTIS>) of which has been integrated into PyView-GUI. This allows interactive visualization of imaging data and inspecting individual image frames or pixelwise time traces. ILTIS can also be used to manually segment images and save the results into files.
 
 # Architecture: Data processing
 The basic idea is to analyze data in two steps: 
