### 2023-01-31 #4936 @ajkswamy
@csoneson Thanks again for your suggestions. We have accordingly revised the manuscript to bring down the word count to ~1600 words. Here is summary of changes:

- the section "Architecture: Data processing" has been drastically shortened by moving content to the wiki
- the sections "Statement of Need" and "Conclusion" have been considerably shortened to remove phrasing and details that were somewhat redundant.

Please have a look and let me know if any further changes are needed to the manuscript, the code repo or its wiki. Thanks!
