### 2023-02-03 #4936 @csoneson
Thanks @ajkswamy - this looks good! I made a [PR](https://github.com/galizia-lab/pyview/pull/35) with some small modifications, please check and merge if you agree.

The next steps then are:

- [ ] Make a tagged release of the software, and list the version tag of the archived version here.
- [ ] Check that the archival deposit (e.g., in Zenodo) has the correct metadata. In particular, the title and author list should be identical to those of the paper. You can also add the authors' ORCID.
