### 2023-02-06 #4936 @ajkswamy
Hi @csoneson. Thanks for the PR with improvements. We have accepted it and made a new release: [v1.3](https://github.com/galizia-lab/pyview/releases/tag/1.3).

The same version of the code has been archived with G-Node Infrastructure services (GIN): https://doi.org/10.12751/g-node.ezh5w2
